feat: 去除侧边栏布局 + 自定义取色器组件(常用色多排+透明度+HEX输入)

+++ backend/internal/controller/engine_status.go
@@ -113,13 +113,13 @@ func (ec *EngineController) HealthCheckPlugin(c *gin.Context) {
 	})
 }
 
-func getTenantIDFromContext(c *gin.Context) uint {
+func getTenantIDFromContext(c *gin.Context) int64 {
 	if tid, exists := c.Get("tenant_id"); exists {
-		if id, ok := tid.(uint); ok {
+		if id, ok := tid.(int64); ok {
 			return id
 		}
 		if id, ok := tid.(float64); ok {
-			return uint(id)
+			return int64(id)
 		}
 	}
 	return 0
